feat(验证码): 1、tianai-captcha升级至1.5.1  2、基于tianai1.5.1重构验证码模块

+++ smart-boot/smart-framework/smart-commons/smart-commons-core/src/main/java/com/smart/framework/commons/core/captcha/dto/CaptchaValidateParameter.java
@@ -74,11 +74,11 @@ public class CaptchaValidateParameter implements Serializable {
         @Serial
         private static final long serialVersionUID = -2301151802413240271L;
         /** x. */
-        private Integer x;
+        private Float x;
         /** y. */
-        private Integer y;
+        private Float y;
         /** 时间. */
-        private Integer t;
+        private Float t;
         /** 类型. */
         private CaptchaTrackTypeEnum type = CaptchaTrackTypeEnum.MOVE;
     }
